+++ src/persist-client/src/cli/inspect.rs
@@ -204,24 +204,9 @@ pub struct StateArgs {
 }
 
 /// Fetches the current state of a given shard
-pub async fn fetch_latest_state(
-    shard_id: ShardId,
-    consensus_uri: &str,
-    blob_uri: &str,
-) -> Result<impl serde::Serialize, anyhow::Error> {
-    let cfg = PersistConfig::new(&READ_ALL_BUILD_INFO, SYSTEM_TIME.clone());
-    let metrics = Arc::new(Metrics::new(&cfg, &MetricsRegistry::new()));
-    let consensus = ConsensusConfig::try_from(
-        consensus_uri,
-        Box::new(cfg.clone()),
-        metrics.postgres_consensus.clone(),
-    )?;
-    let consensus = consensus.clone().open().await?;
-    let blob =
-        BlobConfig::try_from(blob_uri, Box::new(cfg.clone()), metrics.s3_blob.clone()).await?;
-    let blob = blob.clone().open().await?;
-
-    let state_versions = StateVersions::new(cfg, consensus, blob, Arc::clone(&metrics));
+pub async fn fetch_latest_state(args: &StateArgs) -> Result<impl serde::Serialize, anyhow::Error> {
+    let shard_id = args.shard_id();
+    let state_versions = args.open().await?;
     let versions = state_versions
         .fetch_recent_live_diffs::<u64>(&shard_id)
         .await;
@@ -411,7 +396,7 @@ pub async fn blob_batch_part(
 ) -> Result<impl serde::Serialize, anyhow::Error> {
     let cfg = PersistConfig::new(&READ_ALL_BUILD_INFO, SYSTEM_TIME.clone());
     let metrics = Arc::new(Metrics::new(&cfg, &MetricsRegistry::new()));
-    let blob = make_blob(blob_uri, NO_COMMIT, metrics).await?;
+    let blob = make_blob(&cfg, blob_uri, NO_COMMIT, metrics).await?;
 
     let key = PartialBatchKey(partial_key).complete(&shard_id);
     let part = blob
@@ -465,9 +450,8 @@ struct BlobCounts {
 /// Fetches the blob count for given path
 pub async fn blob_counts(blob_uri: &str) -> Result<impl serde::Serialize, anyhow::Error> {
     let cfg = PersistConfig::new(&READ_ALL_BUILD_INFO, SYSTEM_TIME.clone());
-    let metrics = Metrics::new(&cfg, &MetricsRegistry::new());
-    let blob = BlobConfig::try_from(blob_uri, Box::new(cfg), metrics.s3_blob).await?;
-    let blob = blob.clone().open().await?;
+    let metrics = Arc::new(Metrics::new(&cfg, &MetricsRegistry::new()));
+    let blob = make_blob(&cfg, blob_uri, NO_COMMIT, metrics).await?;
 
     let mut blob_counts = BTreeMap::new();
     let () = blob
@@ -772,7 +756,7 @@ impl StateArgs {
         let metrics = Arc::new(Metrics::new(&cfg, &MetricsRegistry::new()));
         let consensus =
             make_consensus(&cfg, &self.consensus_uri, NO_COMMIT, Arc::clone(&metrics)).await?;
-        let blob = make_blob(&self.blob_uri, NO_COMMIT, Arc::clone(&metrics)).await?;
+        let blob = make_blob(&cfg, &self.blob_uri, NO_COMMIT, Arc::clone(&metrics)).await?;
         Ok(StateVersions::new(cfg, consensus, blob, metrics))
     }
 }
